+++ docs/CHANGELOG.md
@@ -12,8 +12,10 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 ### Added
 
 * `amethyst_test` test framework for ergonomic testing of Amethyst applications ([#1000])
+* combinations of buttons triggering actions ([#1043])
 
 [#1000]: https://github.com/amethyst/amethyst/pull/1000
+[#1043]: https://github.com/amethyst/amethyst/pull/1043
 
 ## [0.9.0] - 2018-10